Compilation warning
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@7072 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/bzfs/bzfs.cxx
@@ -3260,7 +3260,7 @@ static void handleCommand(int t, const void *rawbuf)
       player[t].hasSent(message);
       // check for spamming
       std::string tempmsg = message;
-      for (int c = 0; c <= tempmsg.size() - 1; c++)
+      for (int c = 0; c <= (int)tempmsg.size() - 1; c++)
 	if (isspace(tempmsg[c]))
       	  tempmsg.erase(tempmsg.begin() + c);
       if (strcasecmp(tempmsg.c_str(), player[t].getLastMsg().c_str()) == 0 &&
